+++ src/software/thunderscope/BUILD
@@ -72,9 +72,9 @@ py_library(
         "//software/thunderscope/robot_diagnostics:diagnostics_input_widget",
         "//software/thunderscope/robot_diagnostics:drive_and_dribbler_widget",
         "//software/thunderscope/robot_diagnostics:estop_view",
+        "//software/thunderscope/robot_diagnostics:robot_error_log",
         "//software/thunderscope/robot_diagnostics:robot_info",
         "//software/thunderscope/robot_diagnostics:robot_view",
-        "//software/thunderscope/robot_diagnostics/robot_error_log",
     ],
 )
