+++ src/software/thunderscope/constants.py
@@ -1,6 +1,7 @@
 from pyqtgraph.Qt import QtCore, QtGui
 from proto.import_all_protos import *
 from enum import Enum, IntEnum
+from proto.robot_log_msg_pb2 import LogLevel
 
 import textwrap
 
@@ -75,6 +76,18 @@ BALL_HEIGHT_EFFECT_MULTIPLIER = 3
 # Time (in seconds) to sleep / delay the loop iteration for different protos
 # in robot communications
 ROBOT_COMMUNICATIONS_TIMEOUT_S = 0.02
+
+ROBOT_FATAL_TIMEOUT_S = 5
+ROBOT_CRASH_TIMEOUT_S = 5
+
+# LogLevel to string conversion map
+LOG_LEVEL_STR_MAP = {
+    LogLevel.DEBUG: "DEBUG",
+    LogLevel.INFO: "INFO",
+    LogLevel.WARNING: "WARNING",
+    LogLevel.FATAL: "FATAL",
+    LogLevel.CONTRACT: "CONTRACT",
+}
 
 GAME_CONTROLLER_URL = "http://localhost:8081"
