+++ src/software/thunderscope/log/g3log_widget.py
@@ -1,14 +1,12 @@
 import pyqtgraph as pg
 from pyqtgraph.Qt.QtWidgets import *
 import queue
-import time
 from software.py_constants import *
 import pyqtgraph.console as pg_console
 from proto.robot_log_msg_pb2 import RobotLog, LogLevel
 from proto.import_all_protos import *
 
 import software.thunderscope.constants as constants
-from software.networking.threaded_unix_listener import ThreadedUnixListener
 from software.thunderscope.log.g3log_checkboxes import g3logCheckboxes
 from software.thunderscope.thread_safe_buffer import ThreadSafeBuffer
 
@@ -55,16 +53,6 @@ class g3logWidget(QWidget):
         self.robot_last_fatal_time_s = []
         for id in range(MAX_ROBOT_IDS_PER_SIDE):
             self.robot_last_fatal_time_s.append(0)
-        self.ROBOT_FATAL_TIMEOUT_S = 5
-
-        # LogLevel to string conversion map
-        self.log_level_str_map = {
-            LogLevel.DEBUG: "DEBUG",
-            LogLevel.INFO: "INFO",
-            LogLevel.WARNING: "WARNING",
-            LogLevel.FATAL: "FATAL",
-            LogLevel.CONTRACT: "CONTRACT",
-        }
 
     def refresh(self):
         """Update the log widget with another log message
@@ -98,16 +86,7 @@ class g3logWidget(QWidget):
                 and self.checkbox_widget.fatal_checkbox.isChecked()
             )
         ):
-            log_str = f"R{log.robot_id} {log.created_timestamp.epoch_timestamp_seconds} {self.log_level_str_map[log.log_level]} [{log.file_name}->{log.line_number}] {log.log_msg}\n"
+            log_str = f"R{log.robot_id} {log.created_timestamp.epoch_timestamp_seconds} {constants.LOG_LEVEL_STR_MAP[log.log_level]} [{log.file_name}->{log.line_number}] {log.log_msg}\n"
             self.console_widget.repl.write(log_str, style="output")
-            if log.log_level == LogLevel.FATAL or log.log_level == LogLevel.CONTRACT:
-                if (
-                    time.time() - self.robot_last_fatal_time_s[log.robot_id]
-                    > self.ROBOT_FATAL_TIMEOUT_S
-                ):
-                    QMessageBox.information(
-                        self, "Fatal Log Alert", log_str,
-                    )
-                self.robot_last_fatal_time_s[log.robot_id] = time.time()
         else:
             return
