+++ src/software/thunderscope/robot_diagnostics/BUILD
@@ -40,7 +40,6 @@ py_library(
         ":motor_fault_view",
         "//software/thunderscope:constants",
         "//software/thunderscope:thread_safe_buffer",
-        "//software/thunderscope/robot_diagnostics/robot_error_log:error_log_constants",
         requirement("pyqtgraph"),
     ],
 )
@@ -62,7 +61,6 @@ py_library(
         "//software/thunderscope:constants",
         "//software/thunderscope:thread_safe_buffer",
         "//software/thunderscope/robot_diagnostics:robot_info",
-        "//software/thunderscope/robot_diagnostics/robot_error_log:error_log_constants",
         requirement("pyqtgraph"),
     ],
 )
@@ -80,4 +78,23 @@ py_library(
 py_library(
     name = "robot_status",
     srcs = ["robot_status.py"],
+)
+
+py_library(
+    name = "robot_error_log",
+    srcs = ["robot_error_log.py"],
+    deps = [
+        ":error_log_widgets",
+        requirement("pyqtgraph"),
+    ],
+)
+
+py_library(
+    name = "error_log_widgets",
+    srcs = ["error_log_widgets.py"],
+    deps = [
+        ":robot_status",
+        "//software/thunderscope/robot_diagnostics/robot_error_log_icons:error_log_constants",
+        requirement("pyqtgraph"),
+    ],
 )
