+++ src/software/thunderscope/robot_diagnostics/robot_error_log_icons/BUILD
@@ -1,0 +1,14 @@
+package(default_visibility = ["//visibility:public"])
+
+load("@thunderscope_deps//:requirements.bzl", "requirement")
+
+py_library(
+    name = "error_log_constants",
+    srcs = ["error_log_constants.py"],
+    data = [
+        "crash.png",
+        "error_code.png",
+        "fatal_log.png",
+        "low_battery.png",
+    ],
+)
