+++ src/software/thunderscope/robot_diagnostics/robot_error_log_icons/error_code.png
