+++ src/software/thunderscope/robot_diagnostics/robot_error_log_icons/low_battery.png
